Move 'ask_for_serial_port' to splitflap Python module
